+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 5.0.0
+
+* Migrate to 3.29.0
+* Add WebHtmlElementStrategy for ExtendedNetworkImageProvider on Web
+
+## 4.0.6
+
+* Make the package WASM compatible
+
 ## 4.0.5
 
 * Loosen `web` version to `0.3.0~9.x.x`, and fit flutter sdk: >= 3.16